requirements.txt
Flask==3.0.3
line-bot-sdk==3.14.1
openai==1.54.3